requirements.txt
streamlit==1.41.1
snowflake-connector-python==3.12.4
requests==2.32.3
webvtt-py==0.5.1